backend/requirements.txt
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
supabase==2.3.0
